+++ backend/src/main/java/com/clothsell/module/mall/dal/dataobject/cart/CartRespDTO.java
@@ -14,6 +14,7 @@ public class CartRespDTO {
     private String color;
     private String size;
     private BigDecimal price;
+    private BigDecimal amount;
     private Integer stock;
     private Integer qty;
 }