+++ backend/src/main/java/com/clothsell/module/mall/dal/dataobject/product/ProductDO.java
@@ -1,6 +1,5 @@
 package com.clothsell.module.mall.dal.dataobject.product;
 
-import com.baomidou.mybatisplus.annotation.KeySequence;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.clothsell.framework.mybatis.core.dataobject.BaseDO;
@@ -8,7 +7,6 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 @TableName("mall_product")
-@KeySequence("mall_product_seq")
 @Data
 @EqualsAndHashCode(callSuper = true)
 public class ProductDO extends BaseDO {