+++ backend/src/main/java/com/clothsell/module/mall/service/cart/CartService.java
@@ -1,10 +1,8 @@
 package com.clothsell.module.mall.service.cart;
 
-import com.clothsell.module.mall.dal.dataobject.cart.CartRespDTO;
+import com.clothsell.module.mall.dal.dataobject.cart.CartListDTO;
 import com.clothsell.module.mall.vo.cart.CartSaveReqVO;
 import jakarta.validation.Valid;
-
-import java.util.List;
 
 public interface CartService {
     Long createCart(@Valid CartSaveReqVO createReqVO);
@@ -13,5 +11,5 @@ public interface CartService {
 
     void deleteCart(Long id, Long userId);
 
-    List<CartRespDTO> getCartList(Long userId);
+    CartListDTO getCartList(Long userId);
 }